refactor: MostViewedRight as an Island, not a Portal
