+++ dotcom-rendering/src/web/components/MostViewedRight.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 
 import { useApi as useApi_ } from '@root/src/web/lib/useApi';
 
-import { mockTab1 } from '../MostViewed.mocks';
+import { mockTab1 } from './MostViewed/MostViewed.mocks';
 import { MostViewedRight } from './MostViewedRight';
 
 const response = { data: mockTab1 };
