+++ dotcom-rendering/src/web/components/MostViewedRightWrapper.importable.tsx
@@ -11,8 +11,14 @@ type Props = {
 // Minimum height needed to render MostViewedRight is its own outer height.
 const HEIGHT_REQUIRED = 482 + 24 + 24;
 
-const flexGrow = css`
-	flex-grow: 1;
+const MOSTVIEWED_STICKY_HEIGHT = 1059;
+
+// Styling the data island root so it stretches to cover the full height available in the container.
+// Requires us to subtract the height of its sibling in the container (StickyAd).
+const stretchWrapperHeight = css`
+	display: flex;
+	flex-direction: column;
+	height: ${`calc(100% - ${MOSTVIEWED_STICKY_HEIGHT}px)`};
 `;
 
 // Wrapping MostViewedRight so we can determine whether or not there's enough vertical space in the container to render it.
@@ -43,7 +49,7 @@ export const MostViewedRightWrapper = ({ limitItems, isAdFreeUser }: Props) => {
 	}, [heightIsAvailable]);
 
 	return (
-		<div ref={bodyRef} css={flexGrow}>
+		<div ref={bodyRef} css={stretchWrapperHeight}>
 			{heightIsAvailable ? (
 				<MostViewedRight
 					limitItems={limitItems}
